# iter: introduce Iterator with configurable concurrency