Adds configurable concurrency to `iter` via a new type `Iterator` with ~no overhead and no API changes:

```go
customIterator := iter.Iterator{MaxGoroutines: 42}
customIterator.ForEach(...)

// zero value is safe, defaults to runtime.GOMAXPROCS(0)
iter.Iterator{}.ForEach(...)

// package-level functions behave the same as before
iter.ForEach(...)
```

If the API is agreeable, I can make a follow-up PR to introduce something similar for `Map` and friends. Note that `Map` cannot simply be a method on `Iterator` because methods cannot accept type parameters, and we do not want the return type `R` on `Iterator` where `ForEach` does not use `R`.

Relevant to https://github.com/sourcegraph/conc/issues/36, closes https://github.com/sourcegraph/conc/issues/55

---

Before:

```
go test -run=XXX -bench=. -benchmem ./iter
goos: darwin
goarch: arm64
pkg: github.com/sourcegraph/conc/iter
BenchmarkForEach/0-10   13291378                77.89 ns/op          120 B/op          4 allocs/op
BenchmarkForEach/1-10    2048739               607.5 ns/op           144 B/op          5 allocs/op
BenchmarkForEach/8-10     441230              2843 ns/op             312 B/op         12 allocs/op
BenchmarkForEach/100-10                   278910              4847 ns/op             360 B/op         14 allocs/op
BenchmarkForEach/1000-10                   42801             31772 ns/op             360 B/op         14 allocs/op
BenchmarkForEach/10000-10                   2192            521428 ns/op             397 B/op         14 allocs/op
BenchmarkForEach/100000-10                   187           6115313 ns/op            4653 B/op         14 allocs/op
PASS
ok      github.com/sourcegraph/conc/iter        10.492s
```

After:

```
go test -run=XXX -bench=. -benchmem ./iter
goos: darwin
goarch: arm64
pkg: github.com/sourcegraph/conc/iter
BenchmarkForEach/0-10   13504274                77.28 ns/op          120 B/op          4 allocs/op
BenchmarkForEach/1-10    2133229               560.1 ns/op           144 B/op          5 allocs/op
BenchmarkForEach/8-10     488190              2274 ns/op             312 B/op         12 allocs/op
BenchmarkForEach/100-10                   295342              3770 ns/op             360 B/op         14 allocs/op
BenchmarkForEach/1000-10                   33836             35229 ns/op             360 B/op         14 allocs/op
BenchmarkForEach/10000-10                   2001            618155 ns/op             402 B/op         14 allocs/op
BenchmarkForEach/100000-10                   175           6827000 ns/op            4948 B/op         14 allocs/op
PASS
ok      github.com/sourcegraph/conc/iter        10.241s
```